Detach ValueError to a new file from data/value/mod.rs
